handle-takeover-state
